Split verifier and repairer into multiple files
