+++ src/integrity/error.rs
